chatbot-server/requirements.txt
# 核心框架
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic[email]>=2.11.0

# 資料庫
pymongo>=4.13.0
motor>=3.7.0
beanie>=2.0.0

# LangChain 相關
langchain>=0.3.27
langchain-openai>=0.3.28
langchain-community>=0.3.27
langchain-core>=0.3.72
langchain-text-splitters>=0.3.9
langchain-chroma>=0.2.5

# 向量資料庫
chromadb>=1.0.15

# AI 服務
openai>=1.99.0
tiktoken>=0.10.0

# 工具和工具
python-dotenv>=1.1.0
email-validator>=2.2.0

# 文件處理 (可選)
nltk>=3.9.0
pdf2image>=1.16.0
pdfminer.six>=20240706
unstructured>=0.16.0

# 部署
gunicorn>=20.1.0
websockets>=11.0

# 其他工具
tabulate>=0.9.0
six>=1.16.0